Add kernel actions and driver handlers for managing event listeners.
Event namespaces will be known to the kernel but will not normally be
known by the drivers. Changes to an event's listeners need to pass
through the kernel so that a possible namespace may be encoded into the
name before the driver sees it.

The kernel will also need knowledge of the listeners for replication and
to check listened events for validity.

References #3095.